+++ types/pxr-oneline/package.json
@@ -1,14 +1,13 @@
 {
   "private": true,
-  "name": "@types/oneline",
+  "name": "@types/pxr-oneline",
   "version": "1.0.9999",
-  "nonNpm": true,
   "nonNpmDescription": "The OneLine handles all the logic needed for a publisher to serve ads.",
   "projects": [
       "https://github.com/diogopxrnextday/pxr-advertising"
   ],
   "devDependencies": {
-    "@types/oneline": "workspace:."
+    "@types/pxr-oneline": "workspace:."
 },
   "owners": [
       {
